+++ commons/commons-tsv-util/build.gradle.kts
@@ -1,0 +1,9 @@
+import TestDependencies.BaseTestCompileDependencies
+import TestDependencies.BaseTestRuntimeDependencies
+
+dependencies {
+    implementation(kotlin("stdlib"))
+
+    BaseTestCompileDependencies.forEach { testImplementation(it) }
+    BaseTestRuntimeDependencies.forEach { testImplementation(it) }
+}
