+++ commons/commons-tsv-util/src/main/kotlin/ebi/ac/uk/dsl/tsv/TsvDsl.kt
@@ -1,4 +1,4 @@
-package ebi.ac.uk.dsl
+package ebi.ac.uk.dsl.tsv
 
 class Tsv(val lines: MutableList<TsvLine> = mutableListOf()) {
