+++ commons/commons-util/src/test/kotlin/ebi/ac/uk/util/file/ExcelReaderTest.kt
@@ -1,8 +1,8 @@
 package ebi.ac.uk.util.file
 
 import ebi.ac.uk.dsl.excel.excel
-import ebi.ac.uk.dsl.line
-import ebi.ac.uk.dsl.tsv
+import ebi.ac.uk.dsl.tsv.line
+import ebi.ac.uk.dsl.tsv.tsv
 import io.github.glytching.junit.extension.folder.TemporaryFolder
 import io.github.glytching.junit.extension.folder.TemporaryFolderExtension
 import org.assertj.core.api.Assertions.assertThat
