+++ submission/submission-webapp/src/test/kotlin/ac/uk/ebi/biostd/stats/web/handlers/StatsFileHandlerTest.kt
@@ -3,8 +3,8 @@ package ac.uk.ebi.biostd.stats.web.handlers
 import ac.uk.ebi.biostd.persistence.common.model.SubmissionStat
 import ac.uk.ebi.biostd.persistence.common.model.SubmissionStatType.VIEWS
 import ac.uk.ebi.biostd.stats.web.exceptions.InvalidStatException
-import ebi.ac.uk.dsl.line
-import ebi.ac.uk.dsl.tsv
+import ebi.ac.uk.dsl.tsv.line
+import ebi.ac.uk.dsl.tsv.tsv
 import ebi.ac.uk.test.createFile
 import ebi.ac.uk.util.collections.second
 import io.github.glytching.junit.extension.folder.TemporaryFolder
